+++ pages/_app.js
@@ -1,25 +1,15 @@
 import GlobalStyle from "../styles";
 import Head from "next/head";
-import useSWR from "swr";
-
-const fetcher = (url) => fetch(url).then((res) => res.json());
+import Heading from "../components/Heading";
 
 export default function App({ Component, pageProps }) {
-  const {
-    data: poses,
-    error,
-    isLoading,
-  } = useSWR("https://yoga-api-nzy4.onrender.com/v1/poses", fetcher);
-
-  if (error) return <h1>Failed to load</h1>;
-  if (isLoading) return <h1>Loading ...</h1>;
-
   return (
     <>
       <GlobalStyle />
       <Head>
         <title>Capstone Project</title>
       </Head>
+      <Heading />
       <Component {...pageProps} poses={poses} />
     </>
   );
